Fix default environment
I figured there is a problem with pip installs within the conda environment files. That's why I created a dedicated env for Calliope and its pip install. The default env now has no pip installs.

The problem is the following: pip upgrades or downgrades packages that have been installed with conda. That may be convenient, but it's surprising. In this specific case, I declared pandas 0.25 within the conda env, but the actual env contained 0.24. That's because Calliope requires < 0.25 and so pip downgraded pandas.